fix: Program Restructuring
